myapp-gcp-secret-manager-in-application/requirements.txt
tornado
psycopg2
google-cloud-secret-manager
